wkl - NE - product card screen market storage quantity

+++ work_list/src/main/java/com/lenta/bp14/models/work_list/WorkListTask.kt
@@ -235,7 +235,7 @@ data class GoodOptions(
 data class Stock(
         val number: Int,
         val storage: String,
-        val quantity: Int
+        val quantity: Double
 )
 
 data class Provider(
